+++ weblate/api/views.py
@@ -112,11 +112,7 @@ from weblate.trans.models import (
     Unit,
 )
 from weblate.trans.models.translation import Translation, TranslationQuerySet
-from weblate.trans.tasks import (
-    category_removal,
-    component_removal,
-    project_removal,
-)
+from weblate.trans.tasks import category_removal, component_removal, project_removal
 from weblate.trans.views.files import download_multi
 from weblate.trans.views.reports import generate_credits
 from weblate.utils.celery import get_task_progress
@@ -142,9 +138,7 @@ if TYPE_CHECKING:
     from django.http import HttpResponse
     from rest_framework.request import Request
 
-    from weblate.api.serializers import (
-        NewUnitSerializer,
-    )
+    from weblate.api.serializers import NewUnitSerializer
     from weblate.auth.models import AuthenticatedHttpRequest
 
 REPO_OPERATIONS: dict[str, tuple[str, str, tuple, dict, bool]] = {
@@ -572,7 +566,7 @@ class UserViewSet(viewsets.ModelViewSet):
     def notifications(self, request: Request, username: str):
         obj = self.get_object()
         if request.method == "POST":
-            self.perm_check(request, obj)
+            self.perm_check(request, obj, allow_self=True)
             with transaction.atomic():
                 serializer = NotificationSerializer(
                     data=request.data, context={"request": request}
@@ -622,7 +616,7 @@ class UserViewSet(viewsets.ModelViewSet):
             raise Http404(str(error)) from error
 
         if request.method == "DELETE":
-            self.perm_check(request, obj)
+            self.perm_check(request, obj, allow_self=True)
             subscription.delete()
             return Response(status=HTTP_204_NO_CONTENT)
 
@@ -632,7 +626,7 @@ class UserViewSet(viewsets.ModelViewSet):
                 subscription, context={"request": request}
             )
         else:
-            self.perm_check(request, obj)
+            self.perm_check(request, obj, allow_self=True)
             serializer = NotificationSerializer(
                 subscription,
                 data=request.data,
@@ -714,7 +708,11 @@ class GroupViewSet(viewsets.ModelViewSet):
         if user.has_perm("group.edit") or user.has_perm("group.view"):
             queryset = Group.objects.all()
         else:
-            queryset = Group.objects.filter(Q(user=user) | Q(admins=user)).distinct()
+            queryset = Group.objects.filter(
+                Q(user=user)
+                | Q(admins=user)
+                | Q(defining_project__in=user.projects_with_perm("project.permissions"))
+            ).distinct()
         return queryset.order_by("id")
 
     def perm_check(